Added files via upload
Hey Mike.
Here's a bit of code I've been struggling with. Not sure if you can help or not? It involves SESSIONS which would require me sending over the entire login sequence ... a bunch of stuff.
Here's the gist. You can also see a commented out note in the login-prcss_static_pgs.php file.
When a member logs in, I'm trying to query the member's default genre (mmbr_genre inside mmbr_dtl table) to redirect the member's login to their default genre page. Presently, it's hard coded to always land on Folk-Americana page. This should default to the member's preference instead.
I've managed success on the navbar but not in this login form. As you can see, I've added to the SELECT a LEFT JOIN. This brings in the mmbr_dtl table.
I think part of the problem is delineating the two tables and relating the login mmbrname. The two tables relate to each other via the mmbr_key. Some how I need to bring together the logged in members mmbrname and align to their mmbr_key to relate to the mmbr_genre